services/memae-anomaly/requirements.lock.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.lock.txt --python-version 3.11 --python-platform linux
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   starlette
    #   watchfiles
click==8.3.1
    # via uvicorn
fastapi==0.115.8
    # via -r requirements.in
filelock==3.25.2
    # via
    #   torch
    #   triton
fsspec==2026.3.0
    # via torch
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via anyio
jinja2==3.1.6
    # via torch
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.1.1
    # via
    #   -r requirements.in
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
pillow==12.1.1
    # via torchvision
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.2.2
    # via uvicorn
pyyaml==6.0.3
    # via uvicorn
starlette==0.45.3
    # via fastapi
sympy==1.13.1
    # via torch
torch==2.5.1
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.20.1
    # via -r requirements.in
triton==3.1.0
    # via torch
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
uvicorn==0.34.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
